Added a CMAKE file in utils to help compilation.  Performed some renaming
