+++ OGSPlugins/_utils/OGS.cpp
@@ -26,10 +26,10 @@
 #include <omp.h>
 #endif
 
-#include "OGS.hpp"
+#include "OGS.h"
 
 #ifndef OGS_NO_NETCDF
-#include "netcdfio.hpp"
+#include "netcdfio.h"
 #endif
 
 #define CLLIND(ii,jj,kk,nx,ny) ( (nx-1)*(ny-1)*(kk) + (nx-1)*(jj) + (ii) )
