+++ OGSPlugins/_utils/fieldOperations.h
