+++ OGSPlugins/_utils/vtkOperations.cpp
@@ -33,7 +33,7 @@
 #define OMP_THREAD_NUM  0
 #endif
 
-#include "vtkOperations.hpp"
+#include "vtkOperations.h"
 
 #define CLLIND(ii,jj,kk,nx,ny) ( (nx-1)*(ny-1)*(kk) + (nx-1)*(jj) + (ii) )
 #define PNTIND(ii,jj,kk,nx,ny) ( (nx)*(ny)*(kk) + (nx)*(jj) + (ii) )
